Moved testing DB_URL intialization code into own file
