+++ models/init_test.go
@@ -1,0 +1,18 @@
+package models
+
+import (
+	"flag"
+	"os"
+)
+
+var dburl string
+
+// Grab postgres url
+func init() {
+	flag.StringVar(&dburl, "DB_URL", "", "database url for testing")
+	flag.Parse()
+
+	if dburl == "" {
+		dburl = os.Getenv("DB_URL")
+	}
+}
